Move JCEF Shutdown code to a separate class
